requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.0
django-environ==0.11.2
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-registration-redux==2.13
djangorestframework==3.14.0
Pillow==10.1.0
pip==23.3.1
python-decouple==3.8
pytz==2023.3.post1
setuptools==68.2.2
signals==0.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.2
